fix: fix comments after code review
