+++ src/model/cart/cart-model.ts
@@ -39,15 +39,4 @@ class CartModel {
   }
 }
 
-export function formatPrice(price: number): string {
-  const formatPrice = new Intl.NumberFormat('en-US', {
-    style: 'currency',
-    currency: 'USD',
-    minimumFractionDigits: 2,
-    maximumFractionDigits: 2,
-  }).format(price / 100);
-
-  return formatPrice;
-}
-
 export default CartModel;
