+++ src/shared/models/interfaces/handler.ts
@@ -1,0 +1,3 @@
+export interface Handlers {
+  onClick?: (event: Event) => void;
+}
