+++ src/shared/utils/format-price.ts
@@ -1,0 +1,10 @@
+export function formatPrice(price: number): string {
+  const formatPrice = new Intl.NumberFormat('en-US', {
+    style: 'currency',
+    currency: 'USD',
+    minimumFractionDigits: 2,
+    maximumFractionDigits: 2,
+  }).format(price / 100);
+
+  return formatPrice;
+}
